Extract CoolThingDetails
